# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the Voxlab public speaking educational platform",
        "title": "Voxlab API",
        "termsOfService": "http://voxlab.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@voxlab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticates with email + password and returns a JWT token.\nThe token must be sent as `Authorization: Bearer \u003ctoken\u003e` for protected endpoints.\nTokens expire after 24 hours.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login exitoso — token + datos del usuario",
                        "schema": {
                            "$ref": "#/definitions/resources.LoginResponseData"
                        }
                    },
                    "400": {
                        "description": "Credenciales inválidas o formato incorrecto",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Email o contraseña incorrectos",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidates the current JWT token by adding it to a Redis blacklist.\nAfter calling this, the token can no longer be used for authenticated requests.\nThe frontend should also discard the token locally.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Logout",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente",
                        "schema": {
                            "$ref": "#/definitions/resources.LogoutResponse"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's profile (name, email, XP, streak, lives).\nUse this to verify the token is valid and load the user's data on page refresh.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get Current User",
                "responses": {
                    "200": {
                        "description": "Perfil del usuario autenticado",
                        "schema": {
                            "$ref": "#/definitions/resources.UserProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Error al obtener el perfil",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's extended profile (name, email, avatar_url, XP, streak, lives).\nUnlike /auth/me, this returns a richer profile object with all user-facing fields.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "Perfil extendido del usuario",
                        "schema": {
                            "$ref": "#/definitions/resources.UserProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Error al obtener el perfil",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the authenticated user's profile fields (name, avatar_url, etc.).\nSend only the fields you want to change.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Profile fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil actualizado correctamente",
                        "schema": {
                            "$ref": "#/definitions/resources.UserProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos en la solicitud",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el perfil",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Creates a new user account and returns a JWT token (auto-login).\nThe password must be at least 6 characters.\nIf the email is already registered, returns a 409 conflict error.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Registration",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuario creado — auto-login con token",
                        "schema": {
                            "$ref": "#/definitions/resources.RegisterResponseData"
                        }
                    },
                    "400": {
                        "description": "Datos de registro inválidos",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "409": {
                        "description": "El email ya está registrado",
                        "schema": {
                            "$ref": "#/definitions/resources.ConflictError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "resources.BadRequestError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Solicitud inválida — verifique los campos enviados"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.ConflictError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "El recurso ya existe — conflicto con datos existentes"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error interno del servidor"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.LoginResponseData": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2026-01-02T00:00:00Z"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZXhwIjoxNzA2NzU4NDAwfQ.signature"
                },
                "user": {
                    "$ref": "#/definitions/services.UserData"
                }
            }
        },
        "resources.LogoutResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "resources.NotFoundError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Recurso no encontrado"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.RegisterResponseData": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2026-01-02T00:00:00Z"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZXhwIjoxNzA2NzU4NDAwfQ.signature"
                },
                "user": {
                    "$ref": "#/definitions/services.UserData"
                }
            }
        },
...
        "resources.UnauthorizedError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Token de autenticación no proporcionado o inválido"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.UserProfileResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/services.UserData"
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "services.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "services.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                }
            }
        },
        "services.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Jane Doe"
                }
            }
        },
...
        "services.UserData": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://storage.voxlab.com/avatars/abc123.webp"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "streak_days": {
                    "type": "integer",
                    "example": 7
                },
                "xp": {
                    "type": "integer",
                    "example": 1250
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}